requirements.txt
streamlit
pymupdf
python-docx
dotenv
langchain
langchain_google_genai
langchain_openai


